new RSP workflow compatible with shape files
Depends on the output of a new actel function (loadShape) and an updated version of transitionLayer that are not yet available online. Tomorrow I will run the required package tests on them and push them to the repo if everything is checking out.

This also means RSP can now work both with WGS84 coordinates or other coordinate sets. The information of which coordinate set is being used is brought in through the transition layer.

Ran some checks on the runRSP output using the checkLocations function and it seems decent. There was a zone with weird lines that we should probably look into at some point.

There are some important changes on the handling of the base raster that we should discuss too.

ping @yuriniella